chore: yeet react-scripts
