+++ packages/client/.gitignore
@@ -11,6 +11,7 @@
 # production
 /build
 /dist
+/public/devtools.js
 
 # misc
 .DS_Store
