+++ packages/client/package.json
@@ -1,8 +1,17 @@
 {
 	"name": "podium-dev-tools",
 	"version": "0.2.1",
+	"type": "module",
 	"private": true,
 	"homepage": "https://podium-lib.io/",
+	"scripts": {
+		"build": "npm run build:js && npm run build:web-ext",
+		"build:js": "rollup --config rollup.config.js",
+		"build:web-ext": "web-ext build --source-dir public --artifacts-dir dist --overwrite-dest",
+		"start": "npm-run-all --parallel watch:js watch:web-ext",
+		"watch:js": "rollup --config rollup.config.js --watch --watch.onEnd=\"web\"",
+		"watch:web-ext": "web-ext run --source-dir public --devtools"
+	},
 	"dependencies": {
 		"classnames": "2.5.1",
 		"compare-versions": "6.1.0",
@@ -10,35 +19,14 @@
 		"react-dom": "18.2.0"
 	},
 	"devDependencies": {
-		"@testing-library/jest-dom": "6.0.0",
-		"@testing-library/react": "14.0.0",
-		"@testing-library/user-event": "13.5.0",
+		"@babel/core": "7.24.3",
+		"@babel/preset-react": "7.24.1",
+		"@rollup/plugin-babel": "6.0.4",
+		"@rollup/plugin-commonjs": "25.0.7",
+		"@rollup/plugin-node-resolve": "15.2.3",
+		"@rollup/plugin-replace": "5.0.5",
 		"npm-run-all": "4.1.5",
-		"react-scripts": "5.0.1",
+		"rollup": "4.13.1",
 		"web-ext": "7.11.0"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"watch:react": "react-scripts start",
-		"build": "INLINE_RUNTIME_CHUNK=false react-scripts build && web-ext build --source-dir build --artifacts-dir dist --overwrite-dest",
-		"eject": "react-scripts eject"
-	},
-	"eslintConfig": {
-		"extends": [
-			"react-app",
-			"react-app/jest"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
 	}
 }
