+++ packages/client/public/devtools.css
@@ -1,0 +1,55 @@
+body {
+	margin: 0;
+	font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+		"Droid Sans", "Helvetica Neue", sans-serif;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+}
+
+code {
+	font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+}
+
+body {
+	margin: 0px;
+}
+
+.card-body {
+	margin: 0;
+	padding: 3px;
+}
+
+.connection {
+	margin-top: 10px;
+}
+
+.status .meta {
+	font-style: italic;
+	font-size: 0.5em;
+}
+
+.card-header {
+	border-bottom: 0px;
+}
+
+.podlet-name-tabs .nav-item button {
+	margin: 5px;
+	cursor: pointer;
+}
+
+.podlet-name-tabs .nav-item button {
+	margin: 5px;
+	cursor: pointer;
+}
+
+.mb-3 {
+	margin-bottom: 0.1rem !important;
+}
+
+.input-group-text {
+	width: 150px;
+}
+
+.context-form {
+	margin-top: 1px;
+}
