--- packages/client/src/components/app/App.css
@@ -1,8 +1,0 @@
-body {
-    margin: 0px;
-}
-
-.card-body {
-    margin: 0;
-    padding: 3px;
-}
