--- packages/client/src/components/app/connection/Connection.css
@@ -1,3 +1,0 @@
-.connection {
-    margin-top: 10px;
-}
