--- packages/client/src/components/app/container/panel/context/Context.css
@@ -1,16 +1,0 @@
-.podlet-name-tabs .nav-item button {
-    margin: 5px;
-    cursor: pointer;
-}
-
-.mb-3 {
-    margin-bottom: 0.1rem !important;
-}
-
-.input-group-text {
-    width: 150px;
-}
-
-.context-form {
-    margin-top: 1px;
-}
