--- packages/client/src/components/app/container/tabs/tab/Tab.css
@@ -1,4 +1,0 @@
-.podlet-name-tabs .nav-item button {
-    margin: 5px;
-    cursor: pointer;
-}
