--- packages/client/src/components/app/status/Status.css
@@ -1,8 +1,0 @@
-.status .meta {
-    font-style: italic;
-    font-size: 0.5em;
-}
-
-.card-header {
-    border-bottom: 0px;
-}
